Jwt login

+++ controllers/userController.js
@@ -1,6 +1,6 @@
 const userTemplateCopy = require("../schema/userSchema");
 const {
-  isAllValid,
+  isInValid,
   checkUser,
   addUserService,
 } = require("../services/userServices");
@@ -30,6 +30,12 @@ const getUserById = async (req, res) => {
 };
 
 const deleteUser = async (req, res) => {
+  let logUser = req.user._doc;
+
+  // Check for Admin or Corresponding user
+  if (!(logUser._id === req.params.id || logUser.emp_code === "1000"))
+    return res.status(401).json({ msg: "Unauthorized" });
+
   if (await checkUser(req.params.id)) {
     userTemplateCopy
       .findByIdAndRemove(req.params.id)
@@ -45,6 +51,11 @@ const deleteUser = async (req, res) => {
 };
 
 const updateUser = async (req, res) => {
+  let logUser = req.user._doc;
+
+  // Check for Admin or Corresponding user
+  if (!(logUser._id === req.params.id || logUser.emp_code === "1000"))
+    return res.status(401).json({ msg: "Unauthorized" });
   if (await checkUser(req.params.id)) {
     try {
       const user = await userTemplateCopy.findById(req.params.id);
@@ -66,11 +77,18 @@ const updateUser = async (req, res) => {
   }
 };
 
-const upsetUser = async (req, res) => {
-  if (isAllValid(req.body)) {
+const upsertUser = async (req, res) => {
+  let logUser = req.user._doc;
+
+  // Check for Admin or Corresponding user
+  if (!(logUser._id === req.params.id || logUser.emp_code === "1000"))
+    return res.status(401).json({ msg: "Unauthorized" });
+
+  if (isInValid(req.body)) {
     res.status(400).json({ msg: "Bad request" });
     return;
   }
+
   try {
     if (await checkUser(req.params.id)) {
       const user = await userTemplateCopy.findById(req.params.id);
@@ -79,12 +97,12 @@ const upsetUser = async (req, res) => {
 
       user.name = name ? name : user.name;
       user.mobile_no = mobile_no ? mobile_no : user.mobile_no;
-      user.email = email ? email : user.email;
+      // user.email = email ? email : user.email;
 
       const updatedData = await user.save();
 
       res.status(200).json({
-        msg: "Data updated! NOTE: Password and Employee Code can't be updated, contact ADMIN in such case",
+        msg: "Data updated! NOTE: Email, Password and Employee Code can't be updated, contact ADMIN in such case",
         data: updatedData,
       });
     } else {
@@ -96,25 +114,10 @@ const upsetUser = async (req, res) => {
   }
 };
 
-const addUser = async (req, res) => {
-  const userData = req.body;
-  if (isAllValid(userData)) {
-    res.status(400).json({ msg: "Bad request" });
-    return;
-  }
-  try {
-    const resData = await addUserService(userData);
-    res.status(201).json({ msg: "User added successfully", data: resData });
-  } catch (err) {
-    res.status(400).json({ msg: "Err: " + err });
-  }
-};
-
 module.exports = {
   getUsers,
-  addUser,
   getUserById,
   deleteUser,
   updateUser,
-  upsetUser,
+  upsertUser,
 };
